Moved: 'normalize' feature code fully into its own folder.
